Separate MasteryCheckResult from MasteryCheckStatus file
