+++ src/main/java/jarvis/logic/commands/MasteryCheckCommand.java
@@ -10,7 +10,7 @@ import java.util.List;
 import jarvis.commons.core.Messages;
 import jarvis.commons.core.index.Index;
 import jarvis.logic.commands.exceptions.CommandException;
-import jarvis.model.MasteryCheckStatus;
+import jarvis.model.MasteryCheckResult;
 import jarvis.model.Model;
 import jarvis.model.Student;
 
@@ -27,13 +27,13 @@ public class MasteryCheckCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1 " + PREFIX_MC_NUM + "1 " + PREFIX_MC_RES + "PASS";
 
     private final Index index;
-    private final MasteryCheckStatus.MasteryCheckResult mcResult;
+    private final MasteryCheckResult mcResult;
 
     /**
      * @param index of the student in the filtered list
      * @param mcResult result of the student's mastery check
      */
-    public MasteryCheckCommand(Index index, MasteryCheckStatus.MasteryCheckResult mcResult) {
+    public MasteryCheckCommand(Index index, MasteryCheckResult mcResult) {
         this.index = index;
         this.mcResult = mcResult;
     }
